refactor: store
