+++ src/app/providers/storeProvider.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 import { Provider } from "react-redux";
 
-import { store } from "@/shared/store";
+import { store } from "../store";
 
 type Props = {
   children: ReactNode;
