+++ src/app/store/index.ts
@@ -1,0 +1,1 @@
+export { store } from "./store";
