+++ src/app/store/store.ts
@@ -1,0 +1,11 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+import { countriesReducer } from "@/features/countriesAutocomplete";
+import { formReducer } from "@/features/submitHistory";
+
+export const store = configureStore({
+  reducer: {
+    forms: formReducer,
+    countries: countriesReducer,
+  },
+});
