+++ src/shared/store/hooks.ts
@@ -1,0 +1,6 @@
+import { useDispatch, useSelector } from "react-redux";
+
+import { AppDispatch, RootState } from "./types";
+
+export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
+export const useAppSelector = useSelector.withTypes<RootState>();
