--- src/shared/store/store.ts
@@ -1,20 +1,0 @@
-import { useDispatch, useSelector } from "react-redux";
-import { configureStore } from "@reduxjs/toolkit";
-
-import { countriesReducer } from "@/features/countriesAutocomplete";
-import { formReducer } from "@/features/submitHistory";
-
-export const store = configureStore({
-  reducer: {
-    forms: formReducer,
-    countries: countriesReducer,
-  },
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-
-export type AppDispatch = typeof store.dispatch;
-export type AppStore = typeof store;
-
-export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
-export const useAppSelector = useSelector.withTypes<RootState>();
