+++ src/shared/store/types.ts
@@ -1,0 +1,6 @@
+import { store } from "@/app/store";
+
+export type RootState = ReturnType<typeof store.getState>;
+
+export type AppDispatch = typeof store.dispatch;
+export type AppStore = typeof store;
